update to apache.httpcomponents.core5 update musicbrainz.coverart code to client5

+++ src/main/java/net/pms/newgui/update/AutoUpdateDialog.java
@@ -31,9 +31,9 @@ import net.pms.Messages;
 import net.pms.PMS;
 import net.pms.configuration.Build;
 import net.pms.configuration.UmsConfiguration;
+import net.pms.external.update.AutoUpdater;
+import net.pms.external.update.AutoUpdater.State;
 import net.pms.platform.PlatformUtils;
-import net.pms.update.AutoUpdater;
-import net.pms.update.AutoUpdater.State;
 import net.pms.util.FileUtil;
 
 public class AutoUpdateDialog extends JDialog implements Observer {
@@ -174,8 +174,8 @@ public class AutoUpdateDialog extends JDialog implements Observer {
 
 		if (state == State.DOWNLOAD_IN_PROGRESS) {
 			downloadProgressBar.setEnabled(true);
-			downloadProgressBar.setValue(autoUpdater.getBytesDownloaded());
-			downloadProgressBar.setMaximum(autoUpdater.getTotalBytes());
+			downloadProgressBar.setValue(Long.valueOf(autoUpdater.getBytesDownloaded()).intValue());
+			downloadProgressBar.setMaximum(Long.valueOf(autoUpdater.getTotalBytes()).intValue());
 		} else {
 			downloadProgressBar.setEnabled(false);
 			downloadProgressBar.setValue(0);
